requirements.txt
pandas
requests
sqlalchemy
psycopg2-binary
python-dotenv
plotly
dash
fastf1
pyarrow